+++ src/reuse/vcs.py
@@ -1,6 +1,7 @@
 # SPDX-FileCopyrightText: 2017 Free Software Foundation Europe e.V. <https://fsfe.org>
 # SPDX-FileCopyrightText: © 2020 Liferay, Inc. <https://liferay.com>
 # SPDX-FileCopyrightText: 2020 John Mulligan <jmulligan@redhat.com>
+# SPDX-FileCopyrightText: 2023 Markus Haug <korrat@proton.me>
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
@@ -11,10 +12,11 @@ from __future__ import annotations
 import logging
 import os
 from abc import ABC, abstractmethod
+from inspect import isclass
 from pathlib import Path
-from typing import TYPE_CHECKING, Optional, Set
+from typing import TYPE_CHECKING, Generator, Optional, Set, Type
 
-from ._util import GIT_EXE, HG_EXE, StrPath, execute_command
+from ._util import GIT_EXE, HG_EXE, PIJUL_EXE, StrPath, execute_command
 
 if TYPE_CHECKING:
     from .project import Project
@@ -24,6 +26,8 @@ _LOGGER = logging.getLogger(__name__)
 
 class VCSStrategy(ABC):
     """Strategy pattern for version control systems."""
+
+    EXE: str | None = None
 
     @abstractmethod
     def __init__(self, project: Project):
@@ -82,9 +86,11 @@ class VCSStrategyNone(VCSStrategy):
 class VCSStrategyGit(VCSStrategy):
     """Strategy that is used for Git."""
 
+    EXE = GIT_EXE
+
     def __init__(self, project: Project):
         super().__init__(project)
-        if not GIT_EXE:
+        if not self.EXE:
             raise FileNotFoundError("Could not find binary for Git")
         self._all_ignored_files = self._find_all_ignored_files()
         self._submodules = self._find_submodules()
@@ -94,7 +100,7 @@ class VCSStrategyGit(VCSStrategy):
         ignored, don't return all files inside of it.
         """
         command = [
-            str(GIT_EXE),
+            str(self.EXE),
             "ls-files",
             "--exclude-standard",
             "--ignored",
@@ -113,7 +119,7 @@ class VCSStrategyGit(VCSStrategy):
 
     def _find_submodules(self) -> Set[Path]:
         command = [
-            str(GIT_EXE),
+            str(self.EXE),
             "config",
             "-z",
             "--file",
@@ -150,7 +156,7 @@ class VCSStrategyGit(VCSStrategy):
         if not Path(directory).is_dir():
             raise NotADirectoryError()
 
-        command = [str(GIT_EXE), "status"]
+        command = [str(cls.EXE), "status"]
         result = execute_command(command, _LOGGER, cwd=directory)
 
         return not result.returncode
@@ -163,7 +169,7 @@ class VCSStrategyGit(VCSStrategy):
         if not Path(cwd).is_dir():
             raise NotADirectoryError()
 
-        command = [str(GIT_EXE), "rev-parse", "--show-toplevel"]
+        command = [str(cls.EXE), "rev-parse", "--show-toplevel"]
         result = execute_command(command, _LOGGER, cwd=cwd)
 
         if not result.returncode:
@@ -176,9 +182,11 @@ class VCSStrategyGit(VCSStrategy):
 class VCSStrategyHg(VCSStrategy):
     """Strategy that is used for Mercurial."""
 
+    EXE = HG_EXE
+
     def __init__(self, project: Project):
         super().__init__(project)
-        if not HG_EXE:
+        if not self.EXE:
             raise FileNotFoundError("Could not find binary for Mercurial")
         self._all_ignored_files = self._find_all_ignored_files()
 
@@ -187,7 +195,7 @@ class VCSStrategyHg(VCSStrategy):
         is ignored, don't return all files inside of it.
         """
         command = [
-            str(HG_EXE),
+            str(self.EXE),
             "status",
             "--ignored",
             # terse is marked 'experimental' in the hg help but is documented
@@ -218,7 +226,7 @@ class VCSStrategyHg(VCSStrategy):
         if not Path(directory).is_dir():
             raise NotADirectoryError()
 
-        command = [str(HG_EXE), "root"]
+        command = [str(cls.EXE), "root"]
         result = execute_command(command, _LOGGER, cwd=directory)
 
         return not result.returncode
@@ -231,7 +239,7 @@ class VCSStrategyHg(VCSStrategy):
         if not Path(cwd).is_dir():
             raise NotADirectoryError()
 
-        command = [str(HG_EXE), "root"]
+        command = [str(cls.EXE), "root"]
         result = execute_command(command, _LOGGER, cwd=cwd)
 
         if not result.returncode:
@@ -241,6 +249,81 @@ class VCSStrategyHg(VCSStrategy):
         return None
 
 
+class VCSStrategyPijul(VCSStrategy):
+    """Strategy that is used for Pijul."""
+
+    EXE = PIJUL_EXE
+
+    def __init__(self, project: Project):
+        super().__init__(project)
+        if not self.EXE:
+            raise FileNotFoundError("Could not find binary for Pijul")
+        self._all_tracked_files = self._find_all_tracked_files()
+
+    def _find_all_tracked_files(self) -> Set[Path]:
+        """Return a set of all files tracked by pijul."""
+        command = [str(self.EXE), "list"]
+        result = execute_command(command, _LOGGER, cwd=self.project.root)
+        all_files = result.stdout.decode("utf-8").splitlines()
+        return {Path(file_) for file_ in all_files}
+
+    def is_ignored(self, path: StrPath) -> bool:
+        path = self.project.relative_from_root(path)
+        return path not in self._all_tracked_files
+
+    def is_submodule(self, path: StrPath) -> bool:
+        # not supported in pijul yet
+        return False
+
+    @classmethod
+    def in_repo(cls, directory: StrPath) -> bool:
+        if directory is None:
+            directory = Path.cwd()
+
+        if not Path(directory).is_dir():
+            raise NotADirectoryError()
+
+        command = [str(cls.EXE), "diff", "--short"]
+        result = execute_command(command, _LOGGER, cwd=directory)
+
+        return not result.returncode
+
+    @classmethod
+    def find_root(cls, cwd: Optional[StrPath] = None) -> Optional[Path]:
+        if cwd is None:
+            cwd = Path.cwd()
+
+        # TODO this duplicates pijul's logic.
+        # Maybe it should be replaced by calling pijul,
+        # but there is no matching subcommand yet.
+        path = Path(cwd).resolve()
+
+        if not path.is_dir():
+            raise NotADirectoryError()
+
+        while True:
+            if (path / ".pijul").is_dir():
+                return path
+
+            parent = path.parent
+            if parent == path:
+                # We reached the filesystem root
+                return None
+
+            path = parent
+
+
+def all_vcs_strategies() -> Generator[Type[VCSStrategy], None, None]:
+    """Yield all VCSStrategy classes that aren't the abstract base class."""
+    for value in globals().values():
+        if (
+            isclass(value)
+            and issubclass(value, VCSStrategy)
+            and value is not VCSStrategy
+        ):
+            yield value
+
+
 def find_root(cwd: Optional[StrPath] = None) -> Optional[Path]:
     """Try to find the root of the project from *cwd*. If none is found,
     return None.
@@ -248,12 +331,9 @@ def find_root(cwd: Optional[StrPath] = None) -> Optional[Path]:
     Raises:
         NotADirectoryError: if directory is not a directory.
     """
-    if GIT_EXE:
-        root = VCSStrategyGit.find_root(cwd=cwd)
-        if root:
-            return root
-    if HG_EXE:
-        root = VCSStrategyHg.find_root(cwd=cwd)
-        if root:
-            return root
+    for strategy in all_vcs_strategies():
+        if strategy.EXE:
+            root = strategy.find_root(cwd=cwd)
+            if root:
+                return root
     return None