+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 <!--
 SPDX-FileCopyrightText: 2017 Free Software Foundation Europe e.V. <https://fsfe.org>
-SPDX-FileCopyrightText: © 2020 Liferay, Inc. <https://liferay.com>
 SPDX-FileCopyrightText: 2023 DB Systel GmbH
+SPDX-FileCopyrightText: © 2020 Liferay, Inc. <https://liferay.com>
 
 SPDX-License-Identifier: CC-BY-SA-4.0
 -->
@@ -45,9 +45,24 @@ CLI command and its behaviour. There are no guarantees of stability for the
 - Implement handling LicenseRef in `download` and `init`. (#697)
 - Declared support for Python 3.12. (#846)
 - More file types are recognised:
+  - TCL (`.tcl`) (#871)
   - Julia (`.jl`) (#815)
   - Modern Fortran (`.f90`) (#836)
+  - Bazel (`.bzl`) (#870)
+  - GNU Linker script (`.ld`) (#862)
+  - Assembly code (`.s`) (#862)
+  - Empty placeholders (`.empty`) (#862)
+  - ShellCheck configuration (`.shellcheckrc`) (#862)
+  - Pylint in-project configuration (`pylintrc`) (#862)
 - Display recommendations for steps to fix found issues during a lint. (#698)
+- Add support for Pijul VCS. Pijul support is not added to the Docker image.
+  (#858)
+- When running `annotate` on a file with an unrecognised file path, the tool
+  currently exits early. To automatically create a .license file for
+  unrecognised files, `--fallback-dot-license` has been added. (#823, #851,
+  #853, #859; this took a while to get right.)
+- Ignore `.sl` directory as used by [Sapling SCM](https://sapling-scm.com/).
+  (#867)
 
 ### Changed
 
@@ -58,9 +73,6 @@ CLI command and its behaviour. There are no guarantees of stability for the
   "[repositories] that were cloned independently and later added as a submodule
   or old setups", which "have the submodule's git directory inside the submodule
   instead of embedded into the superproject's git directory". (#687)
-- When running `annotate` on a file with an unrecognised file path,
-  automatically create a `.license` file. The old behaviour of exiting on
-  unrecognised paths is now behind `--exit-if-unrecognised`. (#851, #853)
 - No longer scan binary or uncommentable files for their contents in search of
   REUSE information. (#825)
 - `--force-dot-license` and `--skip-unrecognised` are now mutually exclusive on
@@ -80,9 +92,11 @@ CLI command and its behaviour. There are no guarantees of stability for the
 
 ### Fixed
 
+- Syntax errors in .reuse/dep5 now have better error handling. (#841)
 - Reduced python-debian minimum version to 0.1.34. (#808)
 - Fix issue in `annotate` where `--single-line` and `--multi-line` would not
   correctly raise an error with an incompatible comment style. (#853)
+- Fix parsing existing copyright lines when they do not have a year (#861)
 
 ### Security